requirements.txt
# Stock Monitoring System - Python Dependencies
# =============================================
# Install with: pip install -r requirements.txt

# Core Data Libraries
yfinance>=0.2.36          # Yahoo Finance API (free stock data)
pandas>=2.0.0             # Data manipulation
numpy>=1.24.0             # Numerical operations

# Scheduling
schedule>=1.2.0           # Job scheduling
pytz>=2024.1              # Timezone handling

# Email
# (using built-in smtplib and email modules)

# Web Scraping for News
requests>=2.31.0          # HTTP requests
beautifulsoup4>=4.12.0    # HTML parsing
lxml>=5.0.0               # Fast XML/HTML parser

# Configuration
pyyaml>=6.0               # YAML config parsing

# Charts and Visualization (for weekly reports)
matplotlib>=3.8.0         # Chart generation
plotly>=5.18.0            # Interactive charts (optional)

# Date/Time Utilities
python-dateutil>=2.8.2    # Date parsing

# Rate Limiting
ratelimit>=2.2.1          # API rate limiting

# HTML Email Templates
jinja2>=3.1.0             # Template engine

# Logging
colorlog>=6.8.0           # Colored logging output

# Optional: For more reliable news
feedparser>=6.0.0         # RSS feed parsing

# Optional: For earnings calendar
# Note: earnings-calendar package or use yfinance built-in

# Google Trends for sentiment signals (optional, unreliable)
pytrends>=4.9.0           # Google Trends unofficial API

# NYSE Trading Calendar
exchange_calendars>=4.5.0 # Accurate NYSE holiday calendar

# Environment Variables
python-dotenv>=1.0.0      # Load .env files

# Dashboard API Server
flask>=3.0.0              # Web framework
flask-cors>=4.0.0         # CORS support for API

# Development/Testing (optional)
# pytest>=7.0.0
